Refactor. menu.py file to ui folder for more structure and semantics
